+++ commerce/payment/src/main/java/ru/yandex/practicum/payment/service/PaymentService.java
@@ -3,17 +3,18 @@ package ru.yandex.practicum.payment.service;
 import ru.yandex.practicum.interactionapi.dto.OrderDto;
 import ru.yandex.practicum.interactionapi.dto.PaymentDto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface PaymentService {
 
     PaymentDto createPayment(OrderDto orderDto);
 
-    Double getTotalCost(OrderDto orderDto);
+    BigDecimal getTotalCost(OrderDto orderDto);
 
     void paymentRefund(UUID uuid);
 
-    Double productCost(OrderDto orderDto);
+    BigDecimal productCost(OrderDto orderDto);
 
     void paymentFailed(UUID uuid);
 }
